[Chore #161203421] restructure views to separate dir
